Remove 'bitmap' and 'jpeg' as allowed types from the README.
Support for 'bitmap', '.bmp', 'jpeg', and '.jpeg' was removed in
<https://github.com/seamus-oconnor/gulp-jimp/commit/6e796df34fe4e7bbed89550908b79b00674b01ad>,
but the README wasn't updated to reflect the change in behavior.

I also changed the code to require that the given extension be one of
'bmp', 'jpg', or 'png', and to raise an exception if not. I understand
that this could change behavior for some users, but I think this is better
than silently converting to 'png' when a user specifies 'bitmap'.